Refactor: decimal field

+++ order/models.py
@@ -6,7 +6,7 @@ from product.models import Product
 # Create your models here.
 
 class Discount(BaseModel):
-    discount = models.IntegerField(max_digits=2, default=0)
+    discount = models.DecimalField(max_digits=5, decimal_places=2, default=0)
 
     def __str__(self):
         return self.discount
@@ -28,6 +28,6 @@ class Order(BaseModel):
 class OrderItem(BaseModel):
     quantity = models.IntegerField()
     unit_price = models.DecimalField(decimal_places=2, max_digits=5)
-    discount = models.IntegerField(max_digits=2, default=0)
+    discount = models.DecimalField(max_digits=5, decimal_places=2, default=0)
     order = models.ForeignKey(Order, on_delete=models.CASCADE)
     product = models.ForeignKey(Product, on_delete=models.CASCADE)
